feat: implement WebService interface and refactor WebTools to use WebService methods
