+++ build.gradle
@@ -34,7 +34,6 @@ dependencies {
     implementation("com.fasterxml.jackson.core:jackson-databind:2.19.2")
     implementation("org.jsoup:jsoup:1.21.1")
     implementation("org.apache.httpcomponents.client5:httpclient5:5.5")
-    implementation("com.google.cloud:google-cloud-translate:2.72.0")
 
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
