+++ src/test/resources/application.properties
@@ -1,1 +1,2 @@
 allowed.dirs=/tmp
+app.screenshots.api-key=test-api-key
